+++ CHANGELOG.md
@@ -8,6 +8,8 @@ API modifications:
  * lexer.rl: correctly handle __END__ with non-whitespace after it (Peter Zotov)
  * lexer.rl: handle \r in middle of a line as mere whitespace (Peter Zotov)
  * ruby{18,19,20,21}.y, builders/default: precisely point to tUMINUS_NUM. (Peter Zotov)
+ * Parser::Diagnostic: now takes and exposes a symbolic reason (Ian MacLeod)
+ * Parser::ERRORS: moved into errors.rb for better discoverability (Ian MacLeod)
 
 Features implemented:
  * lexer.rl, ruby21.y, builders/default: rational/complex literals. (Peter Zotov)